fix uri and sql

+++ initialize_db.sql
@@ -432,8 +432,8 @@ CREATE TABLE
 CREATE TABLE
   IF NOT EXISTS muted_users (
     id BIGINT AUTO_INCREMENT NOT NULL,
-    user_id BIGINT NOT NULL,
-    muted_user_id BIGINT NOT NULL,
+    user_id INT (11) NOT NULL,
+    muted_user_id INT (11) NOT NULL,
     PRIMARY KEY (id),
     FOREIGN KEY (user_id) REFERENCES metaserver_users (id),
     FOREIGN KEY (muted_user_id) REFERENCES metaserver_users (id)
